Refactor Self Assessment A/B test setup
The set up for the Self Assessment A/B test has been moved into
a concern.

Benefits to this are:
1. The set up doesn't need to added to any new controller
methods

2. It will make it easier to remove when the A/B test has
completed.
